# apps/backend/open-api/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user-onboarding/init-user": {
      "post": {
        "operationId": "UserOnboardingController_initUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitUserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Use Miller BE",
    "description": "Describes the backend api",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InitUserResponseDto": {
        "type": "object",
        "properties": {
          "userWasInitialised": {
            "type": "boolean"
          }
        },
        "required": [
          "userWasInitialised"
        ]
      },
...
    }
  }
}