# libs/shared-api-client/src/models/InitUserResponseDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Use Miller BE
 * Describes the backend api
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface InitUserResponseDto
 */
export interface InitUserResponseDto {
    /**
     * 
     * @type {boolean}
     * @memberof InitUserResponseDto
     */
    userWasInitialised: boolean;
}

export function InitUserResponseDtoFromJSON(json: any): InitUserResponseDto {
    return InitUserResponseDtoFromJSONTyped(json, false);
}

export function InitUserResponseDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): InitUserResponseDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'userWasInitialised': json['userWasInitialised'],
    };
}

export function InitUserResponseDtoToJSON(value?: InitUserResponseDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'userWasInitialised': value.userWasInitialised,
    };
}